Increased timers resolution

+++ core/src/main/scala/pl/tarsa/sortalgobox/Benchmark.scala
@@ -21,5 +21,6 @@
 package pl.tarsa.sortalgobox
 
 trait Benchmark {
-  def forSize(n: Int, validate: Boolean, buffer: Option[Array[Int]] = None): Int
+  def forSize(n: Int, validate: Boolean,
+    buffer: Option[Array[Int]] = None): Long
 }
